requirements.txt
telethon>=1.36.0
pillow>=10.0.0
aiohttp>=3.9.0